同步master代码

+++ kcp.go
@@ -785,9 +785,9 @@ func (kcp *KCP) Update(current uint32) {
 // or optimize ikcp_update when handling massive kcp connections)
 func (kcp *KCP) Check(current uint32) uint32 {
 	ts_flush := kcp.ts_flush
-	tm_flush := 0x7fffffff
-	tm_packet := 0x7fffffff
-	minimal := 0
+	tm_flush := int32(0x7fffffff)
+	tm_packet := int32(0x7fffffff)
+	minimal := uint32(0)
 	if kcp.updated == 0 {
 		return current
 	}
@@ -801,7 +801,7 @@ func (kcp *KCP) Check(current uint32) uint32 {
 		return current
 	}
 
-	tm_flush = int(_itimediff(ts_flush, current))
+	tm_flush = _itimediff(ts_flush, current)
 
 	for k := range kcp.snd_buf {
 		seg := &kcp.snd_buf[k]
@@ -809,28 +809,28 @@ func (kcp *KCP) Check(current uint32) uint32 {
 		if diff <= 0 {
 			return current
 		}
-		if diff < int32(tm_packet) {
-			tm_packet = int(diff)
+		if diff < tm_packet {
+			tm_packet = diff
 		}
 	}
 
-	minimal = int(tm_packet)
+	minimal = uint32(tm_packet)
 	if tm_packet >= tm_flush {
-		minimal = int(tm_flush)
+		minimal = uint32(tm_flush)
 	}
-	if uint32(minimal) >= kcp.interval {
-		minimal = int(kcp.interval)
+	if minimal >= kcp.interval {
+		minimal = kcp.interval
 	}
 
-	return current + uint32(minimal)
+	return current + minimal
 }
 
 // change MTU size, default is 1400
-func (kcp *KCP) SetMtu(mtu int32) int32 {
-	if mtu < 50 || mtu < int32(IKCP_OVERHEAD) {
+func (kcp *KCP) SetMtu(mtu int) int {
+	if mtu < 50 || mtu < IKCP_OVERHEAD {
 		return -1
 	}
-	buffer := make([]byte, (uint32(mtu)+IKCP_OVERHEAD)*3)
+	buffer := make([]byte, (mtu+IKCP_OVERHEAD)*3)
 	if buffer == nil {
 		return -2
 	}
@@ -840,7 +840,7 @@ func (kcp *KCP) SetMtu(mtu int32) int32 {
 	return 0
 }
 
-func (kcp *KCP) Interval(interval int32) int32 {
+func (kcp *KCP) Interval(interval int) int {
 	if interval > 5000 {
 		interval = 5000
 	} else if interval < 10 {
@@ -855,7 +855,7 @@ func (kcp *KCP) Interval(interval int32) int32 {
 // interval: internal update timer interval in millisec, default is 100ms
 // resend: 0:disable fast resend(default), 1:enable fast resend
 // nc: 0:normal congestion control(default), 1:disable congestion control
-func (kcp *KCP) NoDelay(nodelay, interval, resend, nc int32) int32 {
+func (kcp *KCP) NoDelay(nodelay, interval, resend, nc int) int {
 	if nodelay >= 0 {
 		kcp.nodelay = uint32(nodelay)
 		if nodelay != 0 {
@@ -873,16 +873,16 @@ func (kcp *KCP) NoDelay(nodelay, interval, resend, nc int32) int32 {
 		kcp.interval = uint32(interval)
 	}
 	if resend >= 0 {
-		kcp.fastresend = resend
+		kcp.fastresend = int32(resend)
 	}
 	if nc >= 0 {
-		kcp.nocwnd = nc
+		kcp.nocwnd = int32(nc)
 	}
 	return 0
 }
 
 // set maximum window size: sndwnd=32, rcvwnd=32 by default
-func (kcp *KCP) WndSize(sndwnd, rcvwnd int32) int32 {
+func (kcp *KCP) WndSize(sndwnd, rcvwnd int) int {
 	if kcp != nil {
 		if sndwnd > 0 {
 			kcp.snd_wnd = uint32(sndwnd)
@@ -895,6 +895,6 @@ func (kcp *KCP) WndSize(sndwnd, rcvwnd int32) int32 {
 }
 
 // get how many packet is waiting to be sent
-func (kcp *KCP) WaitSnd() int32 {
-	return int32(len(kcp.snd_buf) + len(kcp.snd_queue))
+func (kcp *KCP) WaitSnd() int {
+	return len(kcp.snd_buf) + len(kcp.snd_queue)
 }
